Test if format order changes GHA workflow outcome
Currently the GHA workflow is only resulting in a HTML file being pushed to main, but there should also be a Markdown file too. This change will just test if changing the order of the formats in _quarto.yml leads to a different outcome (like the .md file being pushed to main but not the .html file).